+++ tests/test_core.py
@@ -4,7 +4,6 @@
 """ Unit tests for zeroconf._core """
 
 import asyncio
-import itertools
 import logging
 import os
 import socket
@@ -13,14 +12,13 @@ import threading
 import time
 import unittest
 import unittest.mock
-from typing import Set, cast
-from unittest.mock import MagicMock, patch
+from typing import cast
+from unittest.mock import patch
 
 import pytest
 
 import zeroconf as r
-from zeroconf import NotRunningException, Zeroconf, _core, const, current_time_millis
-from zeroconf._protocol import outgoing
+from zeroconf import NotRunningException, Zeroconf, const, current_time_millis
 from zeroconf.asyncio import AsyncZeroconf
 
 from . import _clear_cache, _inject_response, _wait_for_start, has_working_ipv6
@@ -45,60 +43,6 @@ def threadsafe_query(zc, protocol, *args):
         protocol.handle_query_or_defer(*args)
 
     asyncio.run_coroutine_threadsafe(make_query(), zc.loop).result()
-
-
-# This test uses asyncio because it needs to access the cache directly
-# which is not threadsafe
-@pytest.mark.asyncio
-async def test_reaper():
-    with patch.object(_core, "_CACHE_CLEANUP_INTERVAL", 0.01):
-        aiozc = AsyncZeroconf(interfaces=['127.0.0.1'])
-        zeroconf = aiozc.zeroconf
-        cache = zeroconf.cache
-        original_entries = list(itertools.chain(*(cache.entries_with_name(name) for name in cache.names())))
-        record_with_10s_ttl = r.DNSAddress('a', const._TYPE_SOA, const._CLASS_IN, 10, b'a')
-        record_with_1s_ttl = r.DNSAddress('a', const._TYPE_SOA, const._CLASS_IN, 1, b'b')
-        zeroconf.cache.async_add_records([record_with_10s_ttl, record_with_1s_ttl])
-        question = r.DNSQuestion("_hap._tcp._local.", const._TYPE_PTR, const._CLASS_IN)
-        now = r.current_time_millis()
-        other_known_answers: Set[r.DNSRecord] = {
-            r.DNSPointer(
-                "_hap._tcp.local.",
-                const._TYPE_PTR,
-                const._CLASS_IN,
-                10000,
-                'known-to-other._hap._tcp.local.',
-            )
-        }
-        zeroconf.question_history.add_question_at_time(question, now, other_known_answers)
-        assert zeroconf.question_history.suppresses(question, now, other_known_answers)
-        entries_with_cache = list(itertools.chain(*(cache.entries_with_name(name) for name in cache.names())))
-        await asyncio.sleep(1.2)
-        entries = list(itertools.chain(*(cache.entries_with_name(name) for name in cache.names())))
-        assert zeroconf.cache.get(record_with_1s_ttl) is None
-        await aiozc.async_close()
-        assert not zeroconf.question_history.suppresses(question, now, other_known_answers)
-        assert entries != original_entries
-        assert entries_with_cache != original_entries
-        assert record_with_10s_ttl in entries
-        assert record_with_1s_ttl not in entries
-
-
-@pytest.mark.asyncio
-async def test_reaper_aborts_when_done():
-    """Ensure cache cleanup stops when zeroconf is done."""
-    with patch.object(_core, "_CACHE_CLEANUP_INTERVAL", 0.01):
-        aiozc = AsyncZeroconf(interfaces=['127.0.0.1'])
-        zeroconf = aiozc.zeroconf
-        record_with_10s_ttl = r.DNSAddress('a', const._TYPE_SOA, const._CLASS_IN, 10, b'a')
-        record_with_1s_ttl = r.DNSAddress('a', const._TYPE_SOA, const._CLASS_IN, 1, b'b')
-        zeroconf.cache.async_add_records([record_with_10s_ttl, record_with_1s_ttl])
-        assert zeroconf.cache.get(record_with_10s_ttl) is not None
-        assert zeroconf.cache.get(record_with_1s_ttl) is not None
-        await aiozc.async_close()
-        await asyncio.sleep(1.2)
-        assert zeroconf.cache.get(record_with_10s_ttl) is not None
-        assert zeroconf.cache.get(record_with_1s_ttl) is not None
 
 
 class Framework(unittest.TestCase):
@@ -689,183 +633,6 @@ async def test_multiple_sync_instances_stared_from_async_close():
     assert zc3.loop.is_running()
 
     await asyncio.sleep(0)
-
-
-def test_guard_against_oversized_packets():
-    """Ensure we do not process oversized packets.
-
-    These packets can quickly overwhelm the system.
-    """
-    zc = Zeroconf(interfaces=['127.0.0.1'])
-
-    generated = r.DNSOutgoing(const._FLAGS_QR_RESPONSE)
-
-    for i in range(5000):
-        generated.add_answer_at_time(
-            r.DNSText(
-                "packet{i}.local.",
-                const._TYPE_TXT,
-                const._CLASS_IN | const._CLASS_UNIQUE,
-                500,
-                b'path=/~paulsm/',
-            ),
-            0,
-        )
-
-    try:
-        # We are patching to generate an oversized packet
-        with patch.object(outgoing, "_MAX_MSG_ABSOLUTE", 100000), patch.object(
-            outgoing, "_MAX_MSG_TYPICAL", 100000
-        ):
-            over_sized_packet = generated.packets()[0]
-            assert len(over_sized_packet) > const._MAX_MSG_ABSOLUTE
-    except AttributeError:
-        # cannot patch with cython
-        zc.close()
-        return
-
-    generated = r.DNSOutgoing(const._FLAGS_QR_RESPONSE)
-    okpacket_record = r.DNSText(
-        "okpacket.local.",
-        const._TYPE_TXT,
-        const._CLASS_IN | const._CLASS_UNIQUE,
-        500,
-        b'path=/~paulsm/',
-    )
-
-    generated.add_answer_at_time(
-        okpacket_record,
-        0,
-    )
-    ok_packet = generated.packets()[0]
-
-    # We cannot test though the network interface as some operating systems
-    # will guard against the oversized packet and we won't see it.
-    listener = _core.AsyncListener(zc)
-    listener.transport = unittest.mock.MagicMock()
-
-    listener.datagram_received(ok_packet, ('127.0.0.1', const._MDNS_PORT))
-    assert zc.cache.async_get_unique(okpacket_record) is not None
-
-    listener.datagram_received(over_sized_packet, ('127.0.0.1', const._MDNS_PORT))
-    assert (
-        zc.cache.async_get_unique(
-            r.DNSText(
-                "packet0.local.",
-                const._TYPE_TXT,
-                const._CLASS_IN | const._CLASS_UNIQUE,
-                500,
-                b'path=/~paulsm/',
-            )
-        )
-        is None
-    )
-
-    logging.getLogger('zeroconf').setLevel(logging.INFO)
-
-    listener.datagram_received(over_sized_packet, ('::1', const._MDNS_PORT, 1, 1))
-    assert (
-        zc.cache.async_get_unique(
-            r.DNSText(
-                "packet0.local.",
-                const._TYPE_TXT,
-                const._CLASS_IN | const._CLASS_UNIQUE,
-                500,
-                b'path=/~paulsm/',
-            )
-        )
-        is None
-    )
-
-    zc.close()
-
-
-def test_guard_against_duplicate_packets():
-    """Ensure we do not process duplicate packets.
-    These packets can quickly overwhelm the system.
-    """
-    zc = Zeroconf(interfaces=['127.0.0.1'])
-    listener = _core.AsyncListener(zc)
-    listener.transport = MagicMock()
-
-    query = r.DNSOutgoing(const._FLAGS_QR_QUERY, multicast=True)
-    question = r.DNSQuestion("x._http._tcp.local.", const._TYPE_PTR, const._CLASS_IN)
-    query.add_question(question)
-    packet_with_qm_question = query.packets()[0]
-
-    query3 = r.DNSOutgoing(const._FLAGS_QR_QUERY, multicast=True)
-    question3 = r.DNSQuestion("x._ay._tcp.local.", const._TYPE_PTR, const._CLASS_IN)
-    query3.add_question(question3)
-    packet_with_qm_question2 = query3.packets()[0]
-
-    query2 = r.DNSOutgoing(const._FLAGS_QR_QUERY, multicast=True)
-    question2 = r.DNSQuestion("x._http._tcp.local.", const._TYPE_PTR, const._CLASS_IN)
-    question2.unicast = True
-    query2.add_question(question2)
-    packet_with_qu_question = query2.packets()[0]
-
-    addrs = ("1.2.3.4", 43)
-
-    with patch.object(_core, "current_time_millis") as _current_time_millis, patch.object(
-        listener, "handle_query_or_defer"
-    ) as _handle_query_or_defer:
-        start_time = current_time_millis()
-
-        _current_time_millis.return_value = start_time
-        listener.datagram_received(packet_with_qm_question, addrs)
-        _handle_query_or_defer.assert_called_once()
-        _handle_query_or_defer.reset_mock()
-
-        # Now call with the same packet again and handle_query_or_defer should not fire
-        listener.datagram_received(packet_with_qm_question, addrs)
-        _handle_query_or_defer.assert_not_called()
-        _handle_query_or_defer.reset_mock()
-
-        # Now walk time forward 1000 seconds
-        _current_time_millis.return_value = start_time + 1000
-        # Now call with the same packet again and handle_query_or_defer should fire
-        listener.datagram_received(packet_with_qm_question, addrs)
-        _handle_query_or_defer.assert_called_once()
-        _handle_query_or_defer.reset_mock()
-
-        # Now call with the different packet and handle_query_or_defer should fire
-        listener.datagram_received(packet_with_qm_question2, addrs)
-        _handle_query_or_defer.assert_called_once()
-        _handle_query_or_defer.reset_mock()
-
-        # Now call with the different packet and handle_query_or_defer should fire
-        listener.datagram_received(packet_with_qm_question, addrs)
-        _handle_query_or_defer.assert_called_once()
-        _handle_query_or_defer.reset_mock()
-
-        # Now call with the different packet with qu question and handle_query_or_defer should fire
-        listener.datagram_received(packet_with_qu_question, addrs)
-        _handle_query_or_defer.assert_called_once()
-        _handle_query_or_defer.reset_mock()
-
-        # Now call again with the same packet that has a qu question and handle_query_or_defer should fire
-        listener.datagram_received(packet_with_qu_question, addrs)
-        _handle_query_or_defer.assert_called_once()
-        _handle_query_or_defer.reset_mock()
-
-        log.setLevel(logging.WARNING)
-
-        # Call with the QM packet again
-        listener.datagram_received(packet_with_qm_question, addrs)
-        _handle_query_or_defer.assert_called_once()
-        _handle_query_or_defer.reset_mock()
-
-        # Now call with the same packet again and handle_query_or_defer should not fire
-        listener.datagram_received(packet_with_qm_question, addrs)
-        _handle_query_or_defer.assert_not_called()
-        _handle_query_or_defer.reset_mock()
-
-        # Now call with garbage
-        listener.datagram_received(b'garbage', addrs)
-        _handle_query_or_defer.assert_not_called()
-        _handle_query_or_defer.reset_mock()
-
-    zc.close()
 
 
 def test_shutdown_while_register_in_process():
